Refactor some common methods into a new module & directory
